v0.1.0
Added maps! Admittedly, they're not the best... but they get the job done. Also added menu links at the top, as well as added additional functionality to the location search. You can now look for zip codes AND a place by name - which allows for international searching. I'm sure this will improve in the future.